modify to use f in rest-frame

+++ source/templates.h
@@ -80,7 +80,7 @@ double kappa_ind_comp(PlasmaPtr xplasma, double freq);
 double total_comp(WindPtr one, double t_e);
 double klein_nishina(double nu);
 void set_comp_func_values(double rand_cs, double max_cs, double energy_ratio);
-int compton_dir(PhotPtr p);
+double compton_dir(PhotPtr p);
 double pdf_thermal(double x, void *params);
 int compton_get_thermal_velocity(double t, double *v);
 double compton_func(double f, void *params);
